Rename receiever persist.rs to session.rs
Renamed `persist.rs` to `session.rs` as the file no longer manages
persistence logic. It now handles session replays and history
management.
