+++ payjoin/src/receive/v2/session.rs
